Silence MSVC truncation warning when calculating alternate bucket index.
MSVC warns that when calculating the alternate bucket index the hash (calculated
by multiplying a 64-bit literal value) may get truncated (to a 32-bit
size_t). True, but it shouldn't be a problem. This commit adds an explicit cast
to silence that warning.

+++ src/cuckoohash_map.hh
@@ -922,7 +922,8 @@ private:
         // ensure tag is nonzero for the multiply.
         const partial_t nonzero_tag = (partial >> 1 << 1) + 1;
         // 0xc6a4a7935bd1e995 is the hash constant from 64-bit MurmurHash2
-        const size_t hash_of_tag = nonzero_tag * 0xc6a4a7935bd1e995;
+        const size_t hash_of_tag =
+            static_cast<size_t>(nonzero_tag * 0xc6a4a7935bd1e995);
         return (index ^ hash_of_tag) & hashmask(hp);
     }
